@@ -1,23 +1,24 @@
 <?php
 
-namespace Locastic\ActivityLog\Message;
+namespace Locastic\Loggastic\Message;
 
-use Locastic\ActivityLog\Enum\ActivityLogAction;
+use Locastic\Loggastic\Enum\ActivityLogAction;
+use Locastic\Loggastic\Util\ClassUtils;
 
-class CreateActivityLogMessage
+class CreateActivityLogMessage implements CreateActivityLogMessageInterface
 {
-    private $item;
-    private string $actionName;
-    private \DateTime $dateTime;
+    private readonly string $actionName;
+    private readonly \DateTime $dateTime;
+    private ?array $userInfo = null;
+    private ?string $requestUrl = null;
 
-    public function __construct($item, ?string $actionName = null)
+    public function __construct(private readonly object $item, ?string $actionName = null)
     {
         $this->dateTime = new \DateTime();
-        $this->item = $item;
-        $this->actionName = $actionName ?? ActivityLogAction::$CREATED;
+        $this->actionName = $actionName ?? ActivityLogAction::CREATED;
     }
 
-    public function getItem()
+    public function getItem(): object
     {
         return $this->item;
     }
@@ -30,5 +31,35 @@ class CreateActivityLogMessage
     public function getActionName(): string
     {
         return $this->actionName;
+    }
+
+    public function getUser(): ?array
+    {
+        return $this->userInfo;
+    }
+
+    public function setUser(?array $userInfo): void
+    {
+        $this->userInfo = $userInfo;
+    }
+
+    public function getRequestUrl(): ?string
+    {
+        return $this->requestUrl;
+    }
+
+    public function setRequestUrl(?string $requestUrl): void
+    {
+        $this->requestUrl = $requestUrl;
+    }
+
+    public function getObjectId()
+    {
+        return $this->getItem()->getId();
+    }
+
+    public function getClassName(): string
+    {
+        return ClassUtils::getClass($this->getItem());
     }
 }
