@@ -1,24 +1,16 @@
 <?php
 
-namespace Locastic\ActivityLog\Metadata\LoggableContext\Factory;
+namespace Locastic\Loggastic\Metadata\LoggableContext\Factory;
 
-use ApiPlatform\Core\Util\ReflectionClassRecursiveIterator;
-use Locastic\ActivityLog\Annotation\Loggable;
-use Locastic\ActivityLog\Metadata\LoggableContext\LoggableContextCollection;
-use Doctrine\Common\Annotations\Reader;
+use Locastic\Loggastic\Metadata\LoggableContext\LoggableContextCollection;
 use Symfony\Contracts\Cache\CacheInterface;
 
 class CachedLoggableContextCollectionFactory implements LoggableContextCollectionFactoryInterface
 {
-    public const CACHE_KEY = 'loggable_name_collection';
+    final public const CACHE_KEY = 'loggable_name_collection';
 
-    private CacheInterface $cache;
-    private LoggableContextCollectionFactoryInterface $decorated;
-
-    public function __construct(LoggableContextCollectionFactoryInterface $decorated, CacheInterface $cache)
+    public function __construct(private readonly LoggableContextCollectionFactoryInterface $decorated, private readonly CacheInterface $cache)
     {
-        $this->cache = $cache;
-        $this->decorated = $decorated;
     }
 
     /**
@@ -26,8 +18,6 @@ class CachedLoggableContextCollectionFactory implements LoggableContextCollectio
      */
     public function create(): LoggableContextCollection
     {
-        return $this->cache->get(self::CACHE_KEY, function () {
-            return $this->decorated->create();
-        });
+        return $this->cache->get(self::CACHE_KEY, fn() => $this->decorated->create());
     }
 }
