@@ -1,19 +1,14 @@
 <?php
 
-namespace Locastic\ActivityLog\Metadata\LoggableContext\Factory;
+namespace Locastic\Loggastic\Metadata\LoggableContext\Factory;
 
-use Locastic\ActivityLog\Metadata\Extractor\LoggableExtractorInterface;
-use Locastic\ActivityLog\Metadata\LoggableContext\LoggableContextCollection;
+use Locastic\Loggastic\Metadata\Extractor\LoggableExtractorInterface;
+use Locastic\Loggastic\Metadata\LoggableContext\LoggableContextCollection;
 
 final class ExtractorLoggableContextCollectionFactory implements LoggableContextCollectionFactoryInterface
 {
-    private $extractor;
-    private $decorated;
-
-    public function __construct(LoggableExtractorInterface $extractor, LoggableContextCollectionFactoryInterface $decorated = null)
+    public function __construct(private readonly LoggableExtractorInterface $extractor, private readonly ?\Locastic\Loggastic\Metadata\LoggableContext\Factory\LoggableContextCollectionFactoryInterface $decorated = null)
     {
-        $this->extractor = $extractor;
-        $this->decorated = $decorated;
     }
 
     /**
