@@ -1,17 +1,14 @@
 <?php
 
-namespace Locastic\ActivityLog\Bridge\Elasticsearch;
+namespace Locastic\Loggastic\Bridge\Elasticsearch;
 
 use Elastic\Elasticsearch\Client;
 use Elastic\Elasticsearch\ClientBuilder;
 
 class ElasticsearchClient
 {
-    private string $activityLogElasticHost;
-
-    public function __construct(string $activityLogElasticHost)
+    public function __construct(private readonly string $activityLogElasticHost)
     {
-        $this->activityLogElasticHost = $activityLogElasticHost;
     }
 
     public function getClient(): Client
