@@ -1,59 +1,63 @@
 <?php
 
-namespace Locastic\ActivityLog\Command;
+namespace Locastic\Loggastic\Command;
 
-use Locastic\ActivityLog\Bridge\Elasticsearch\Index\ElasticsearchIndexFactoryInterface;
-use Locastic\ActivityLog\Metadata\LoggableContext\Factory\LoggableContextCollectionFactoryInterface;
+use Elastic\Elasticsearch\Exception\ClientResponseException;
+use Locastic\Loggastic\Bridge\Elasticsearch\Index\ElasticsearchIndexFactoryInterface;
+use Locastic\Loggastic\Metadata\LoggableContext\Factory\LoggableContextCollectionFactoryInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
+#[\Symfony\Component\Console\Attribute\AsCommand('locastic:activity-logs:create-loggable-indexes')]
 class CreateLoggableIndexesCommand extends Command
 {
-    protected static $defaultName = 'locastic:create-loggable-indexes';
-
-    private LoggableContextCollectionFactoryInterface $loggableContextCollectionFactory;
-    private ElasticsearchIndexFactoryInterface $elasticsearchIndexFactory;
-
     protected function configure(): void
     {
         $this->setName(self::$defaultName);
     }
 
-    public function __construct(LoggableContextCollectionFactoryInterface $loggableContextCollectionFactory, ElasticsearchIndexFactoryInterface $elasticsearchIndexFactory)
+    public function __construct(private readonly LoggableContextCollectionFactoryInterface $loggableContextCollectionFactory, private readonly ElasticsearchIndexFactoryInterface $elasticsearchIndexFactory)
     {
         parent::__construct();
-        $this->loggableContextCollectionFactory = $loggableContextCollectionFactory;
-        $this->elasticsearchIndexFactory = $elasticsearchIndexFactory;
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $this->io = new SymfonyStyle($input, $output);
+        $io = new SymfonyStyle($input, $output);
 
-        $this->io->title('Creating activity log indexes ...');
+        $io->title('Creating activity log indexes ...');
         $loggableContextCollection = $this->loggableContextCollectionFactory->create();
 
         foreach ($loggableContextCollection->getIterator() as $loggableClass => $config) {
-            $this->io->writeln('Creating ' . $loggableClass . ' activity_log index');
+            $io->writeln('Creating '.$loggableClass.' activity_log index');
 
             try {
                 $this->elasticsearchIndexFactory->createActivityLogIndex($loggableClass);
             } catch (\Exception $e) {
-                $this->io->error($e->getMessage());
+                if (strpos($e->getMessage(), 'resource_already_exists_exception')) {
+                    $output->writeln('Index already exists, skipping.');
+                } else {
+                    throw $e;
+                }
             }
 
-            $this->io->writeln('Creating ' . $loggableClass . ' current_data_tracker index');
+            $io->writeln('Creating '.$loggableClass.' current_data_tracker index');
 
             try {
                 $this->elasticsearchIndexFactory->createCurrentDataTrackerLogIndex($loggableClass);
-            } catch (\Exception $e) {
-                $this->io->error($e->getMessage());
+            } catch (ClientResponseException $e) {
+                if (strpos($e->getMessage(), 'resource_already_exists_exception')) {
+                    $output->writeln('Index already exists, skipping.');
+                } else {
+                    throw $e;
+                }
             }
         }
 
-        $this->io->success('Done!');
-        return 0;
+        $io->success('Done!');
+
+        return Command::SUCCESS;
     }
 }
