@@ -1,26 +1,21 @@
 <?php
 
-namespace Locastic\ActivityLog\DataProvider;
+namespace Locastic\Loggastic\DataProvider;
 
-use Locastic\ActivityLog\Bridge\Elasticsearch\Context\ElasticsearchContextFactoryInterface;
-use Locastic\ActivityLog\Bridge\Elasticsearch\ElasticsearchService;
-use Locastic\ActivityLog\Model\ActivityLog;
-use Locastic\ActivityLog\Model\CurrentDataTracker;
+use Locastic\Loggastic\Bridge\Elasticsearch\Context\ElasticsearchContextFactoryInterface;
+use Locastic\Loggastic\Bridge\Elasticsearch\ElasticsearchService;
+use Locastic\Loggastic\Model\ActivityLog;
+use Locastic\Loggastic\Model\CurrentDataTracker;
 
 class ActivityLogProvider implements ActivityLogProviderInterface
 {
-    private ElasticsearchService $elasticsearchService;
-    private ElasticsearchContextFactoryInterface $elasticsearchContextFactory;
-
-    public function __construct(ElasticsearchService $elasticsearchService, ElasticsearchContextFactoryInterface $elasticsearchContextFactory)
+    public function __construct(private readonly ElasticsearchService $elasticsearchService, private readonly ElasticsearchContextFactoryInterface $elasticsearchContextFactory)
     {
-        $this->elasticsearchService = $elasticsearchService;
-        $this->elasticsearchContextFactory = $elasticsearchContextFactory;
     }
 
     public function getActivityLogsByClass(string $className, array $sort = []): array
     {
-        $elasticContext = $this->elasticsearchContextFactory->createFromClassName($className);
+        $elasticContext = $this->elasticsearchContextFactory->create($className);
 
         $body = [
             'sort' => $sort,
@@ -36,34 +31,53 @@ class ActivityLogProvider implements ActivityLogProviderInterface
 
     public function getCurrentDataTrackerByClassAndId(string $className, $objectId): ?array
     {
-        $elasticContext = $this->elasticsearchContextFactory->createFromClassName($className);
+        $elasticContext = $this->elasticsearchContextFactory->create($className);
 
         $body = [
-            'query' => ['term' => ['objectId' => $objectId]]
+            'query' => ['term' => ['objectId' => $objectId]],
         ];
 
         // todo move class to config
         return $this->elasticsearchService->getItemByQuery(
             $elasticContext->getCurrentDataTrackerIndex(),
             CurrentDataTracker::class,
-            $body);
+            $body
+        );
     }
 
-    public function getActivityLogsByClassAndId(string $className, $objectId, array $sort = []): array
+    public function getActivityLogsByClassAndId(string $className, mixed $objectId, array $sort = []): array
     {
-        $elasticContext = $this->elasticsearchContextFactory->createFromClassName($className);
+        $elasticContext = $this->elasticsearchContextFactory->create($className);
 
         $body = [
             'sort' => $sort,
             'query' => [
                 'term' => ['objectId' => $objectId],
-            ]
+            ],
         ];
 
         // todo move class to config
         return $this->elasticsearchService->getCollection(
             $elasticContext->getActivityLogIndex(),
             ActivityLog::class,
-            $body);
+            $body
+        );
+    }
+
+    public function getActivityLogsByIndexAndId(string $index, $objectId, array $sort = []): array
+    {
+        $body = [
+            'sort' => $sort,
+            'query' => [
+                'term' => ['objectId' => $objectId],
+            ],
+        ];
+
+        // todo move class to config
+        return $this->elasticsearchService->getCollection(
+            $index,
+            ActivityLog::class,
+            $body
+        );
     }
 }
