@@ -1,8 +1,7 @@
 <?php
 
-namespace Locastic\ActivityLog\Metadata\Extractor;
+namespace Locastic\Loggastic\Metadata\Extractor;
 
-use ApiPlatform\Exception\InvalidArgumentException;
 use Symfony\Component\Yaml\Exception\ParseException;
 use Symfony\Component\Yaml\Yaml;
 
@@ -11,11 +10,11 @@ class YamlLoggableExtractor extends AbstractLoggableExtractor
     protected function extractPath(string $path): void
     {
         try {
-            $loggableClasses = Yaml::parse((string)file_get_contents($path), Yaml::PARSE_CONSTANT);
+            $loggableClasses = Yaml::parse((string) file_get_contents($path), Yaml::PARSE_CONSTANT);
         } catch (ParseException $e) {
             $e->setParsedFile($path);
 
-            throw new InvalidArgumentException($e->getMessage(), $e->getCode(), $e);
+            throw new \InvalidArgumentException($e->getMessage(), $e->getCode(), $e);
         }
 
         if (null === $loggableClasses = $loggableClasses['locastic_loggable'] ?? $loggableClasses) {
@@ -23,7 +22,7 @@ class YamlLoggableExtractor extends AbstractLoggableExtractor
         }
 
         if (!\is_array($loggableClasses)) {
-            throw new InvalidArgumentException(sprintf('"locastic_loggable" setting is expected to be null or an array, %s given in "%s".', \gettype($loggableClasses), $path));
+            throw new \InvalidArgumentException(sprintf('"locastic_loggable" setting is expected to be null or an array, %s given in "%s".', \gettype($loggableClasses), $path));
         }
 
         $this->buildLoggableClasses($loggableClasses, $path);
@@ -40,15 +39,15 @@ class YamlLoggableExtractor extends AbstractLoggableExtractor
     private function validateLoggableClass(array $loggableClassConfig, $path): void
     {
         if (!array_key_exists('class', $loggableClassConfig) || !array_key_exists('groups', $loggableClassConfig)) {
-            throw new InvalidArgumentException(sprintf('Invalid configuration for "locastic_loggable" in "%s".', $path));
+            throw new \InvalidArgumentException(sprintf('Invalid configuration for "locastic_loggable" in "%s".', $path));
         }
 
         if (!is_array($loggableClassConfig['groups'])) {
-            throw new InvalidArgumentException(sprintf('Invalid configuration for "locastic_loggable": "groups" must be an array, %s given in "%s".', \gettype($loggableClass['groups']), $path));
+            throw new \InvalidArgumentException(sprintf('Invalid configuration for "locastic_loggable": "groups" must be an array, %s given in "%s".', \gettype($loggableClassConfig['groups']), $path));
         }
 
         if (!class_exists($loggableClassConfig['class'])) {
-            throw new InvalidArgumentException(sprintf('Invalid configuration for "locastic_loggable" in "%s": class "%s" does not exist.', $path, $loggableClass['class']));
+            throw new \InvalidArgumentException(sprintf('Invalid configuration for "locastic_loggable" in "%s": class "%s" does not exist.', $path, $loggableClassConfig['class']));
         }
     }
 }
