@@ -1,14 +1,11 @@
 <?php
 
-namespace Locastic\ActivityLog\Metadata\LoggableContext\Factory;
+namespace Locastic\Loggastic\Metadata\LoggableContext\Factory;
 
 class LoggableContextFactory implements LoggableContextFactoryInterface
 {
-    private LoggableContextCollectionFactoryInterface $loggableContextCollectionFactory;
-
-    public function __construct(LoggableContextCollectionFactoryInterface $loggableContextCollectionFactory)
+    public function __construct(private readonly LoggableContextCollectionFactoryInterface $loggableContextCollectionFactory)
     {
-        $this->loggableContextCollectionFactory = $loggableContextCollectionFactory;
     }
 
     public function create(string $loggableClass): ?array
